moves lists (genres) to shared story
