+++ client/src/editor/store/initialState.js
@@ -1,7 +1,4 @@
 export default {
   drafts: {},
   publishedAdventures: [],
-  lists: {
-    genres: [],
-  },
 };
