+++ client/src/editor/store/selectors/index.js
@@ -1,6 +1,5 @@
 export { default as draftSelector } from './draftSelector';
 export { default as draftsSelector } from './draftsSelector';
-export { default as genresSelector } from './genresSelector';
 export {
   default as publishedAdventureSelector,
 } from './publishedAdventureSelector';
