+++ client/src/shared/constants/queries.js
@@ -1,0 +1,10 @@
+import { gql } from 'apollo-boost';
+
+export const GET_GENRES = gql`
+  {
+    genres {
+      id
+      name
+    }
+  }
+`;
