+++ client/src/shared/store/selectors/index.js
@@ -1,1 +1,2 @@
+export { default as genresSelector } from './genresSelector';
 export { default as tokenSelector } from './tokenSelector';
